+++ app/src/main/java/com/hga/appturismo/bdSQLite/SqliteLugar.java
@@ -237,7 +237,7 @@ public class SqliteLugar extends DBSQLiteParent implements SqliteInterface<Model
         modeloLugarTuristico.setDescripcion(cursor.getString(cursor.getColumnIndex(DBModel.LUGARES_DESCRIPCION)));
         modeloLugarTuristico.setHorario(cursor.getString(cursor.getColumnIndex(DBModel.LUGARES_HORARIO_ATENCION)));
         modeloLugarTuristico.setDireccion(cursor.getString(cursor.getColumnIndex(DBModel.LUGARES_UBICACION)));
-        modeloLugarTuristico.setTelefono(cursor.getInt(cursor.getColumnIndex(DBModel.LUGARES_TELEFONO)));
+        modeloLugarTuristico.setTelefono(cursor.getLong(cursor.getColumnIndex(DBModel.LUGARES_TELEFONO)));
         modeloLugarTuristico.setGpsX(cursor.getFloat(cursor.getColumnIndex(DBModel.LUGARES_LATITUD)));
         modeloLugarTuristico.setGpsY(cursor.getFloat(cursor.getColumnIndex(DBModel.LUGARES_LONGITUD)));
         modeloLugarTuristico.setActividad(cursor.getString(cursor.getColumnIndex(DBModel.LUGARES_ACTIVIDAD)));
